dev-audio

+++ resources/views/Media/audio/index.blade.php
@@ -7,6 +7,7 @@
 <link href="{{ asset('plugins/iCheck/all.css') }}" rel="stylesheet" type="text/css" />
 <link href="{{ asset('plugins/tagsinput/jquery.tagsinput.min.css') }}" rel="stylesheet" type="text/css" />
 
+
 @endsection
 
 
@@ -32,6 +33,28 @@
     box-sizing: border-box;
     }
 
+    .np-image-upload-picker {
+      padding: 20px;
+      background: #eee;
+      border-radius: 16px;
+      margin: 10px;
+      }
+    np-image-preview {
+      padding: 20px;
+      background: #eee;
+      border-radius: 16px;
+      margin: 10px;
+      }
+
+    img.np-preview {
+        background-color: #fff;
+        /* border: 1px solid #ddd; */
+        padding: 5px;
+        height: 230px;
+        width: 400px;
+        margin: 10px;
+      }
+
 
     .has-spinner .fa-spinner {
     opacity: 0;
@@ -51,6 +74,121 @@
     opacity: 1;
     max-width: 50px; 
     }
+
+    @import url('https://fonts.googleapis.com/css2?family=Lato:wght@300&display=swap');
+
+    .music-container{
+        background-color: #fff;
+        border-radius: 15px;
+        box-shadow: 0 10px 10px 0 grey;
+        display: flex;
+        padding: 10px 20px;
+        position: relative;
+        margin: 100px 0;
+        /* z-index: 10; */
+        width: 400px;
+    }
+    body{
+        font-family: 'Lato', sans-serif;
+    }
+    .img-container{
+       
+        position: relative;
+        width: 110px; 
+    }
+
+    .img-container img {
+        border-radius: 50%;
+        width: inherit;
+        object-fit: cover;
+        height: 110px;
+        position: absolute;
+        bottom: 0;
+        left: 0;
+        animation: rotate 3s linear infinite;
+        animation-play-state: paused;
+    }
+    .music-container.play .img-container img{
+        animation-play-state: running;
+    }
+    
+    @keyframes rotate {
+        from{
+            transform: rotate(0deg)
+        }
+        to{
+            transform: rotate(360deg)
+        }
+    }
+
+
+
+    .navigation{
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        z-index: 1;
+    }
+
+    .action-btn{
+        background-color: #fff;
+        border: 0;
+        font-size: 20px;
+        color: silver;
+        cursor: pointer;
+        padding: 10px;
+        margin: 0 20px;
+    }
+    .action-btn .action-btn-big{
+        color: #cdc2d0;
+        font-size: 30px;
+    }
+    .action-btn:focus{
+        outline: none;
+    }
+
+    .music-info{
+        background-color: rgba(255, 255, 255, 0.5);
+        border-radius: 15px 15px 0 0;
+        position: absolute;
+        opacity: 0;
+        width: calc(100% - 40px);
+        padding: 10px 10px 10px 150px;
+        top: 0;
+        left: 20px;
+        transform: translateY(0%);
+        transition: transform .3s ease-in, opacity .3s ease-in;
+        z-index: 0;
+
+    }
+    .music-info h4{
+        margin: 0;
+    }
+    .progress-container{
+        background-color: #fff;
+        border-radius: 5px;
+        cursor: pointer;
+        margin: 10px 0;
+        height: 5px;
+        width: 100%;
+    }
+
+    .music-container.play .music-info{
+        opacity: 1;
+        transform: translateY(-100%);
+    }
+
+    .progress{
+        height: 100%;
+        border-radius: 5%;
+        width:50%;
+        background-color: green;
+        transition: width .1s linear;
+
+    }
+    
+
+    
 </style>
 
 @endsection
@@ -92,12 +230,25 @@
                 </thead>
                 <tbody>
                     <tr v-for="(audio, index) in audios">
-                        
                         <td v-cloak>
-                            <audio controls>
-                                <!-- <source :src="audio.media_path">     -->
-                                <source :src="'/storage/audios/' + audio.file">    
-                            </audio>
+                            <div class="music-container" id="music-container">
+                                <div class="music-info">
+                                    <h4 id="title">@{{ audio.name }}</h4>
+                                  <div class="progress-container" id="progresscontainer">
+                                    <div id="progress" class="progress"></div>
+                                  </div>
+                                </div>
+                             
+                               <audio id="audio" :src="'/storage/audios/' + audio.file"></audio>
+                               <div class="img-container" v-if="audio.img_cover">
+                                    <img v-cloak id="music-cover" width="20" height="20" :src="'/storage/audio/cover' + audio.img_cover" alt="">
+                               </div>
+                               <div class="navigation">
+                                   <!-- <button id="prev" class="action-btn"><i class="fas fa-backward"></i></button> -->
+                                   <button id="play" class="action-btn action-btn-big"><i class="fas fa-play"></i></button>
+                                   <!-- <button id="next" class="action-btn"><i class="fas fa-forward"></i></button> -->
+                               </div>
+                            </div>
                         </td>
                         <td v-cloak>@{{ audio.name }}</td>
                         <td v-cloak>@{{ audio.category }}</td>
@@ -107,17 +258,20 @@
                                 <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                     <i class="fas fa-th-large"></i> 
                                 </button>
-                                <ul class="dropdown-menu">                      
-                                    <li role="separator" class="divider"></li>
-                                    <li><a href="javascript:void(0)"  data-target="#changeTrack" >Change File</a></li>
-                                    <li role="separator" class="divider"></li>
-                                    <li><a href="javascript:void(0)" @click="showDialogInfo(index)" data-toggle="modal"  data-target="#edit-track">Edit Details </a></li>
-                                    <li role="separator" class="divider"></li>
-                                    <li><a href="javascript:void(0)" @click="downloadTrack(audio.file)" title="Download" class="alert-link">Download</li>
-                                    <li><a href="javascript:void(0)" @click="deletePlaylist(index)" class="alert-link">Delete</a></li>
+                                <ul class="dropdown-menu">    
+                                    @if(Auth::user())                  
+                                        <li role="separator" class="divider"></li>
+                                        <li><a href="javascript:void(0);"  data-target="#changeTrack" >Change File</a></li>
+                                        <li role="separator" class="divider"></li>
+                                        <li><a href="javascript:void(0);" @click="showDialogInfo(index)" data-toggle="modal"  data-target="#edit-track">Edit Details </a></li>
+                                        <li role="separator" class="divider"></li>
+                                        <li><a href="javascript:void(0)" @click="deletePlaylist(index)" class="alert-link">Delete</a></li>
+                                    @endif
+                                    <li>
+                                        <a href="javascript:void(0);" @click="downloadTrack(audio.file)" title="Download" class="alert-link">Download
+                                    </li>
                                 </ul>
-                            </div>
-                          
+                            </div>        
                         </td>
                     </tr>
                 </tbody>
@@ -140,8 +294,8 @@
                     </div>
                     <div class="modal-body">
                         <div class="form-group">
-                        <label for="">Name</label>
-                        <input type="text" class="form-control" placeholder="Give a name" v-model="audio.name">
+                        <label for="">Title</label>
+                        <input type="text" class="form-control" placeholder="Track Title" v-model="audio.title">
                         </div>
                         <div class="form-group">
                             <label for="">Category</label>
@@ -155,6 +309,22 @@
                             <div  class="text-center">
                                 <audio id="audio-preview" style="box-shadow:3px 2px 3px grey;" controls v-show="file != ''"/>
                             </div> 
+                        </div>
+
+                        <div class="form-group">
+                            <div  v-if="imageFile == null || imageFile.length == 0" class="fileUpload px-5 py-2 mt-5 btn-md btn-primary" style="height:20px; color:white; padding: 2px; border-radius:3px;">
+                                 <span> Audio Cover</span>&nbsp;&nbsp;<i v-if="!imageFile"></i>
+                                <input type="file" id="uploadBtn file" name="img_cover" accept="image/*" @change="audioCoverPreview($event)" class="form-control upload">      
+                            </div>
+                            <div class="text-center" v-if="imageFile != null && imageFile.length != 0" >
+                                <img class="np-preview" :src="imageFile" />
+                            </div> 
+
+                            <div v-if="imageFile != null && imageFile.length != 0 && !isImageUploading">
+                                <button class="btn-sm btn-default" v-on:click="clearImage()" title="Remove">
+                                    <i class="fas fa-trash"></i>
+                                </button>
+                            </div>
                         </div>
                         
                     </div>
@@ -186,7 +356,7 @@
                     </div>
                     <div class="modal-body">
                         <div class="form-group">
-                        <label for="">Name</label>
+                        <label for="">Title</label>
                         <input type="text" class="form-control" placeholder="Give a name" v-model="audioEdit.name">
                         </div>
                         <div class="form-group">
@@ -226,6 +396,52 @@
 	<script src="{{ asset('plugins/iCheck/icheck.min.js') }}" type="text/javascript"></script>
 	<script src="{{ asset('plugins/tagsinput/jquery.tagsinput.min.js') }}" type="text/javascript"></script>
 
+  <script>
+      const musicContainer =  document.getElementById("music-coantainer");
+      const progressContainer =  document.getElementById("progress-coantainer");
+      const playBtn =  document.getElementById("play");
+      const prevBtn =  document.getElementById("prev");
+      const nextBtn =  document.getElementById("next");
+      const audio =  document.getElementById("audio");
+
+      const title =  document.getElementById("title");
+
+
+      const songs = [];
+      let songIndex = 2;
+      loadSong(songs[songIndex]);
+
+      function loadSong(song){
+         title.innerText = song; 
+        
+      }
+
+      playBtn.addEventListener('click', () =>{
+          const isPlaying = musicContainer.classList.contains("play");
+          isPlaying ? pauseSong() : playSong();
+        
+      });
+
+      function playSong(){
+          musicContainer.classList.add("play");
+          playBtn.querySelector('i.fas').classList.remove("fa-play");
+          playBtn.querySelector('i.fas').classList.add("fa-pause");
+
+          audio.play();
+      }
+
+
+      function pauseSong(){
+          musicContainer.classList.add("play");
+          playBtn.querySelector('i.fas').classList.remove("fa-play");
+          playBtn.querySelector('i.fas').classList.add("fa-pause");
+
+          audio.pause();
+      }
+
+
+
+  </script>
     <script>
 
         $(function(){
